refactor: split nav-pilot into focused modules
Break the 1022-line monolith main.go into focused files:
- main.go (144 lines) — CLI parsing via testable run() function
- install.go (509 lines) — install/uninstall/status commands
- manifest.go (102 lines) — manifest loading + validation
- files.go (152 lines) — file ops, hashing, conflict detection
- state.go (54 lines) — state file read/write
- source.go (91 lines) — source resolution and git clone
- output.go (29 lines) — color helpers

Key improvements:
- Testable CLI: run() returns errors instead of os.Exit
- Manifest validation: rejects empty names, path traversal, duplicates
- Hash error propagation: no more silent _ on fileHash/dirHash
- countDirFiles now recurses into subdirectories
- timeNow variable for deterministic test timestamps

Added 13 new tests: CLI parsing (8), manifest validation (5).
Coverage: 50.0% → 56.6% (46 tests total).
